+++ packages/edition-slices/src/tiles/shared/index.js
@@ -1,5 +1,6 @@
+import getCrop from "./get-crop";
 import TileImage from "./tile-image";
 import TileLink from "./tile-link";
 import TileSummary from "./tile-summary";
 
-export { TileImage, TileLink, TileSummary };
+export { getCrop, TileImage, TileLink, TileSummary };