+++ packages/article-magazine-standard/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.15.0](https://github.com/newsuk/times-components/compare/@times-components/article-magazine-standard@0.14.4...@times-components/article-magazine-standard@0.15.0) (2019-02-18)
+
+
+### Features
+
+* Article Magazine Standard tablet byline ([#1709](https://github.com/newsuk/times-components/issues/1709)) ([904cd23](https://github.com/newsuk/times-components/commit/904cd23))
+
+
+
+
+
 ## [0.14.4](https://github.com/newsuk/times-components/compare/@times-components/article-magazine-standard@0.14.3...@times-components/article-magazine-standard@0.14.4) (2019-02-18)
 
 **Note:** Version bump only for package @times-components/article-magazine-standard